Reorganise subject models Move all subject code into the SubjectStore directory.
