+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/InteractionLayer/components/TranscribedLines/TranscribedLines.spec.js
@@ -1,10 +1,10 @@
 import { shallow } from 'enzyme'
 import sinon from 'sinon'
 import React from 'react'
-import TranscriptionReductions from '@store/TranscriptionReductions'
+import TranscriptionReductions from '@store/SubjectStore/Subject/TranscriptionReductions'
 import taskRegistry from '@plugins/tasks'
 import { TranscribedLines } from './TranscribedLines'
-import { reducedSubject } from '@store/TranscriptionReductions/mocks'
+import { reducedSubject } from '@store/SubjectStore/Subject/TranscriptionReductions/mocks'
 import { TranscriptionLine } from '@plugins/drawingTools/components'
 import ConsensusPopup from './components/ConsensusPopup'
 import { expect } from 'chai'
