+++ packages/lib-classifier/src/store/SubjectStore/SingleImageSubject/SingleImageSubject.js
