+++ packages/lib-classifier/src/store/SubjectStore/SingleImageSubject/SingleImageSubject.spec.js
@@ -1,7 +1,7 @@
 import sinon from 'sinon'
 import SingleImageSubject from './SingleImageSubject'
-import RootStore from '../'
-import WorkflowStore from '../WorkflowStore'
+import RootStore from '@store/'
+import WorkflowStore from '@store/WorkflowStore'
 import { SubjectFactory, WorkflowFactory } from '@test/factories'
 import stubPanoptesJs from '@test/stubPanoptesJs'
 import subjectViewers from '@helpers/subjectViewers'
