+++ packages/lib-classifier/src/store/SubjectStore/SingleImageSubject/index.js
