+++ packages/lib-classifier/src/store/SubjectStore/Subject/Subject.spec.js
@@ -1,11 +1,11 @@
 import { Factory } from 'rosie'
 import sinon from 'sinon'
 import Subject from './Subject'
-import ProjectStore from '../ProjectStore'
-import WorkflowStore from '../WorkflowStore'
+import ProjectStore from '@store/ProjectStore'
+import WorkflowStore from '@store/WorkflowStore'
 import { ProjectFactory, SubjectFactory, WorkflowFactory } from '@test/factories'
 import stubPanoptesJs from '@test/stubPanoptesJs'
-import RootStore from '../'
+import RootStore from '@store/'
 import subjectViewers from '@helpers/subjectViewers'
 import { subjectsSeenThisSession } from '@helpers'
