+++ packages/lib-classifier/src/store/SubjectStore/Subject/TranscriptionReductions/mocks/reducedASMSubject.json
