+++ packages/lib-classifier/src/store/SubjectStore/Subject/TranscriptionReductions/mocks/reducedEmptySubject.js
