+++ packages/lib-classifier/src/store/SubjectStore/Subject/TranscriptionReductions/mocks/reducedSubject.js
