+++ packages/lib-classifier/src/store/SubjectStore/SubjectStore.spec.js
@@ -4,9 +4,9 @@ import nock from 'nock'
 import { Factory } from 'rosie'
 import sinon from 'sinon'
 
-import RootStore from '../RootStore'
-import SubjectGroup from '../SubjectGroup'
-import SingleImageSubject from '../SingleImageSubject'
+import RootStore from '@store/RootStore'
+import SubjectGroup from './SubjectGroup'
+import SingleImageSubject from './SingleImageSubject'
 import { openTalkPage, MINIMUM_QUEUE_SIZE } from './SubjectStore'
 import { ProjectFactory, SubjectFactory, SubjectSetFactory, WorkflowFactory } from '@test/factories'
 import mockStore from '@test/mockStore'
